Hard refactor

+++ src/main/java/Models/Coords.java
@@ -1,11 +1,11 @@
 package Models;
 
 public class Coords {
-    public int X;
-    public int Y;
-    public int Z;
+    public double X;
+    public double Y;
+    public double Z;
 
-    public Coords(int x, int y, int z) {
+    public Coords(double x, double y, double z) {
         X = x;
         Y = y;
         Z = z;
